Move WorkflowList component from components to view directory
